+++ green/src/main/java/com/blockstream/green/ui/receive/ReceiveFragment.kt
@@ -63,9 +63,8 @@ class ReceiveFragment : WalletFragment<ReceiveFragmentBinding>(
         binding.vm = viewModel
 
         binding.address.setOnClickListener {
-            copyToClipboard("Address", binding.address.text.toString(), requireContext())
+            copyToClipboard("Address", binding.address.text.toString(), requireContext(), animateView = binding.address)
             snackbar(R.string.id_address_copied_to_clipboard)
-            binding.address.pulse()
         }
 
         binding.buttonShare.setOnClickListener {