+++ green/src/main/java/com/blockstream/green/ui/wallet/TransactionDetailsFragment.kt
@@ -2,7 +2,10 @@ package com.blockstream.green.ui.wallet
 
 import android.content.Intent
 import android.os.Bundle
+import android.view.MenuItem
 import android.view.View
+import android.widget.TextView
+import androidx.annotation.MenuRes
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.MutableLiveData
 import androidx.navigation.fragment.navArgs
@@ -11,6 +14,7 @@ import com.blockstream.gdk.data.Transaction
 import com.blockstream.green.R
 import com.blockstream.green.databinding.BaseRecyclerViewBinding
 import com.blockstream.green.databinding.ListItemGenericDetailBinding
+import com.blockstream.green.databinding.ListItemTransactionAmountBinding
 import com.blockstream.green.databinding.ListItemTransactionProgressBinding
 import com.blockstream.green.gdk.getConfirmationsMax
 import com.blockstream.green.ui.WalletFragment
@@ -18,22 +22,22 @@ import com.blockstream.green.ui.items.*
 import com.blockstream.green.ui.looks.TransactionDetailsLook
 import com.blockstream.green.utils.*
 import com.greenaddress.greenbits.ui.send.SendAmountActivity
+import com.kennyc.bottomsheet.BottomSheetListener
+import com.kennyc.bottomsheet.BottomSheetMenuDialogFragment
 import com.mikepenz.fastadapter.FastAdapter
 import com.mikepenz.fastadapter.GenericItem
 import com.mikepenz.fastadapter.adapters.FastItemAdapter
 import com.mikepenz.fastadapter.adapters.GenericFastItemAdapter
 import com.mikepenz.fastadapter.binding.listeners.addClickListener
 import com.mikepenz.fastadapter.diff.FastAdapterDiffUtil
-import com.mikepenz.itemanimators.AlphaCrossFadeAnimator
 import com.mikepenz.itemanimators.AlphaInAnimator
-import com.mikepenz.itemanimators.SlideDownAlphaAnimator
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
 @AndroidEntryPoint
 class TransactionDetailsFragment : WalletFragment<BaseRecyclerViewBinding>(
     layout = R.layout.base_recycler_view,
-    menuRes = 0
+    menuRes = R.menu.menu_transaction_details
 ) {
     override val isAdjustResize: Boolean = true
 
@@ -46,7 +50,7 @@ class TransactionDetailsFragment : WalletFragment<BaseRecyclerViewBinding>(
 
     private val saveButtonEnabled = MutableLiveData(false)
 
-    lateinit var noteListItem : GenericDetailListItem
+    lateinit var noteListItem: GenericDetailListItem
 
     @Inject
     lateinit var viewModelFactory: TransactionDetailsViewModel.AssistedFactory
@@ -75,7 +79,7 @@ class TransactionDetailsFragment : WalletFragment<BaseRecyclerViewBinding>(
             )
         }
 
-        viewModel.editableNote.observe(viewLifecycleOwner){
+        viewModel.editableNote.observe(viewLifecycleOwner) {
             saveButtonEnabled.postValue(it != viewModel.originalNote.value ?: "")
         }
 
@@ -102,20 +106,43 @@ class TransactionDetailsFragment : WalletFragment<BaseRecyclerViewBinding>(
             viewModel.bumpFee()
         }
 
-        fastAdapter.addClickListener<ListItemGenericDetailBinding, GenericItem>({ binding -> binding.button }) { _, i: Int, fastAdapter: FastAdapter<GenericItem>, item: GenericItem ->
-            if(item == noteListItem){
+        fastAdapter.addClickListener<ListItemGenericDetailBinding, GenericItem>({ binding -> binding.button }) { view, i: Int, fastAdapter: FastAdapter<GenericItem>, item: GenericItem ->
+            if (item == noteListItem) {
                 viewModel.saveNote()
                 hideKeyboard()
-            }else{
-                openBrowser("${session.network.explorerUrl}${args.transaction.txHash}")
+            } else {
+                if (session.isLiquid) {
+                    showPopupMenu(view, R.menu.menu_transaction_details_explorer_liquid) { item ->
+                        when (item.itemId) {
+                            R.id.explorer_confidential -> {
+                                openBrowser("${session.network.explorerUrl}${args.transaction.txHash}")
+                            }
+                            R.id.explorer_non_confidential -> {
+                                val blinder = args.transaction.getUnblindedString().let {
+                                    if (it.isNotBlank()) "#blinded=$it" else ""
+                                }
+
+                                openBrowser("${session.network.explorerUrl}${args.transaction.txHash}$blinder")
+                            }
+                        }
+                        true
+                    }
+                } else {
+                    openBrowser("${session.network.explorerUrl}${args.transaction.txHash}")
+                }
             }
         }
 
+        fastAdapter.addClickListener<ListItemTransactionAmountBinding, GenericItem>({ binding -> binding.addressTextView }) { view, i: Int, fastAdapter: FastAdapter<GenericItem>, item: GenericItem ->
+            copyToClipboard("Address", (view as TextView).text.toString(), animateView = view)
+            snackbar(R.string.id_address_copied_to_clipboard)
+        }
+
         fastAdapter.onClickListener = { _, _, item: GenericItem, position: Int ->
-            when(item){
+            when (item) {
                 is TransactionAmountListItem -> {
                     val assetId = item.look.assets[item.index].first
-                    if(assetId != session.policyAsset){
+                    if (assetId != session.policyAsset) {
                         navigate(
                             TransactionDetailsFragmentDirections.actionTransactionDetailsFragmentToAssetBottomSheetFragment(
                                 assetId = assetId,
@@ -136,6 +163,21 @@ class TransactionDetailsFragment : WalletFragment<BaseRecyclerViewBinding>(
         }
     }
 
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
+            R.id.share -> {
+                if (session.isLiquid) {
+                    showMenu(R.menu.menu_transaction_details_share_liquid)
+                } else {
+                    share("${session.network.explorerUrl}${args.transaction.txHash}")
+                }
+
+                return true
+            }
+        }
+        return super.onOptionsItemSelected(item)
+    }
+
     private fun updateAdapter(
         transaction: Transaction,
     ) {
@@ -148,9 +190,7 @@ class TransactionDetailsFragment : WalletFragment<BaseRecyclerViewBinding>(
             list += TransactionAmountListItem(transaction, i, look)
         }
 
-        if (transaction.isOut || transaction.isRedeposit) {
-            list += TransactionFeeListItem(transaction, look)
-        }
+        list += TransactionFeeListItem(transaction, look)
 
         FastAdapterDiffUtil.set(amountsAdapter.itemAdapter, list, true)
 
@@ -165,7 +205,7 @@ class TransactionDetailsFragment : WalletFragment<BaseRecyclerViewBinding>(
 
         list += TitleListItem(title = StringHolder(R.string.id_transaction_details))
         val confirmations = transaction.getConfirmations(session.blockHeight)
-        if(confirmations > session.network.confirmationsRequired && transaction.spv.disabledOrVerified()) {
+        if (confirmations > session.network.confirmationsRequired && transaction.spv.disabledOrVerified()) {
             list += GenericDetailListItem(
                 title = StringHolder(R.string.id_confirmations),
                 content = StringHolder("$confirmations")
@@ -181,6 +221,55 @@ class TransactionDetailsFragment : WalletFragment<BaseRecyclerViewBinding>(
         list += noteListItem
 
         FastAdapterDiffUtil.set(detailsAdapter.itemAdapter, list, true)
+    }
+
+    private fun showMenu(@MenuRes menuRes: Int) {
+        BottomSheetMenuDialogFragment.Builder(
+            context = requireContext(),
+            style = R.style.Green_BottomSheetMenuDialog,
+            sheet = menuRes,
+            listener = object : BottomSheetListener {
+                override fun onSheetDismissed(
+                    bottomSheet: BottomSheetMenuDialogFragment,
+                    `object`: Any?,
+                    dismissEvent: Int
+                ) {
+
+                }
+
+                override fun onSheetItemSelected(
+                    bottomSheet: BottomSheetMenuDialogFragment,
+                    item: MenuItem,
+                    `object`: Any?
+                ) {
+
+                    when (item.itemId) {
+                        R.id.non_confidential_transaction -> {
+                            val blinder = args.transaction.getUnblindedString().let {
+                                if (it.isNotBlank()) "#blinded=$it" else ""
+                            }
+
+                            share("${session.network.explorerUrl}${args.transaction.txHash}$blinder")
+                        }
+                        R.id.confidential_transaction -> {
+                            share("${session.network.explorerUrl}${args.transaction.txHash}")
+                        }
+                        R.id.unblinding_data -> {
+                            share(args.transaction.getUnblindedData().toString())
+                        }
+                    }
+                }
+
+                override fun onSheetShown(
+                    bottomSheet: BottomSheetMenuDialogFragment,
+                    `object`: Any?
+                ) {
+
+                }
+
+            },
+            title = getString(R.string.id_share),
+        ).show(childFragmentManager)
     }
 
     override fun getWalletViewModel() = viewModel